simplify manual test(s)
