--- tests/manual/test_api_integration.py
@@ -1,39 +1,0 @@
-"""
-Manual API integration tests.
-
-These tests interact with real external APIs and require valid API keys.
-They should be run manually and not as part of automated CI/CD.
-
-To run:
-1. Set up .env with real API keys
-2. Run with: pytest tests/manual/test_api_integration.py -v
-"""
-
-import os
-import pytest
-from dotenv import load_dotenv
-from artificial_u.audio.processor import AudioProcessor
-from artificial_u.generators.content import ContentGenerator
-
-
-def setup_module():
-    """Load environment variables for API keys."""
-    load_dotenv()
-
-
-@pytest.mark.manual
-@pytest.mark.skipif(
-    not os.getenv("ELEVENLABS_API_KEY"), reason="Requires ElevenLabs API key"
-)
-def test_elevenlabs_voice_creation():
-    """Test creating a voice with ElevenLabs API."""
-    pass  # Test implementation would go here
-
-
-@pytest.mark.manual
-@pytest.mark.skipif(
-    not os.getenv("ANTHROPIC_API_KEY"), reason="Requires Anthropic API key"
-)
-def test_anthropic_content_generation():
-    """Test generating content with Anthropic Claude API."""
-    pass  # Test implementation would go here
